+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.2]
+
+### Changed
+
+- Removed queue dependency and implemented a static promise map
+- Possibly existing .tmp files will be cleaned during construction
+
+### Added
+
+- Some more tests, verifying external file manipulation synchronization and large writes
+
 ## [1.2.1] - 21 May 2025
 
 ### Fixed